% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{chunk.combo}
\alias{chunk.combo}
\title{chunk.combo}
\usage{
chunk.combo(combo,chunkSize)
}
\arguments{
\item{combo}{an ultraCombo object}

\item{chunkSize}{size of chunks desired}
}
\description{
chunk.combo
}
\examples{
n<-20
k<-10
combo<-ultraCombo(seq(choose(n,k)),n,k)
combo
chunk.combo(combo,5e4)
}

