% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.ultraCombo.R
\name{chunk.ultraCombo}
\alias{chunk.ultraCombo}
\title{chunk.ultraCombo}
\usage{
\method{chunk}{ultraCombo}(combo, chunkSize)
}
\arguments{
\item{combo}{an 'ultraCombo' to chunk}

\item{chunkSize}{maximum number of indices per chunk.}
}
\description{
chunk.ultraCombo
}
