% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combnG.R
\name{combnG}
\alias{combnG}
\title{Combination Generator}
\usage{
combnG(i, n, k, .combine = rbind)
}
\arguments{
\item{i}{A vector of indices}

\item{n}{an integer, n of combination}

\item{k}{an integer, k of combination}

\item{.combine}{the function with which the requested combinations will be bound.}
}
\value{
a 'vector' or 'matrix'
the requested combinations bound as specified.
}
\description{
Generates combinations from a
reference integer vector. combnG(x,n,k)
is equivalent to combn(n,k)[,x] but does not
generate the rest of the table.\cr
}
\examples{
n<-20
k<-10
stopifnot(0==sum(combnG(1,n,k)!=seq(k)))
stopifnot(0==sum(combnG(2,n,k)!=c(seq(k-1),k+1)))
stopifnot(0==sum(combnG(choose(n,k),n,k)!=seq(to=n,length.out=k)))
combnG(ceiling(runif(10)*choose(n,k)),n,k)

if(ultraCombo:::debugFlag==FALSE){
stopifnot(sum(combn(n,k)!=combnG(seq(choose(n,k)),n,k,cbind))==0) 
}

}
\author{
Dale Potter, \email{dale@piratepress.org}
}
