% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comboChunk.R
\name{comboChunk}
\alias{comboChunk}
\title{comboChunk}
\usage{
comboChunk(combo, ...)
}
\arguments{
\item{combo}{an 'ultraCombo' to chunk}

\item{...}{arguments for bit::chunk}
}
\description{
Chunk combos. See bit::chunk
}
