% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{createCombo}
\alias{createCombo}
\title{createCombo}
\usage{
createCombo(indices, n, k)
}
\arguments{
\item{indices}{A vector of combination indices.}
}
\value{
A list with the following elements:\cr
	$indices The indices of the combination set\cr
	$n The n of the combination set.\cr
	$k The k of the combination set.\cr
	$Gen A function which generates combinations from the indices
}
\description{
Create a combination object.
}
\examples{
n1<-10
k1<-5
n2<-choose(n1,k1)
k2<-5
i<-ceiling(runif(1)*choose(n2,k2))
combo<-createCombo(combnG(i,n2,k2),n1,k1)
combo
combo$Gen(seq_along(combo$indices))
}

