% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCombo.R
\name{dataCombo}
\alias{dataCombo}
\title{dataCombo}
\usage{
dataCombo(combo, dataObj, FUN = return)
}
\arguments{
\item{combo}{an 'ultraCombo'}

\item{dataObj}{a data object}

\item{FUN}{a 'function' used to process data after slicing. Default is invisible()}

\item{...}{extra args for '['}
}
\description{
A method to wrap a data object with an ultraCombo.
}
