% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growCombo.R
\name{growCombo}
\alias{growCombo}
\title{growCombo}
\usage{
growCombo(name, nb, k = 7, seeds = 0)
}
\arguments{
\item{name}{a character name of the desired area.}

\item{nb}{a 'nb' object defining element relationships}

\item{k}{a 'numeric' dictating the size in elements required, default is 7.}

\item{seeds}{a numeric vector containing the regions to grow, or if negative, the regions to exclude. Default is 0 (which grows all)}
}
\value{
an 'ultraCombo' containing the expanded regions.
}
\description{
Compute geographically contiguous
regions by looped expansion.
}
