% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growCombo.R
\name{growComboSingle}
\alias{growComboSingle}
\title{growComboSingle}
\usage{
growComboSingle(nb, combo)
}
\arguments{
\item{nb}{a 'nb' object defining element relationships}

\item{combo}{a 'list' object of combination definitions.}
}
\value{
a 'numeric' vector of combination indices for k+1
}
\description{
Expand a set of geographically contiguous regions by one element.
}
