% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superChoose.R
\name{hashChoose}
\alias{hashChoose}
\title{hashChoose}
\usage{
hashChoose(n, k)
}
\arguments{
\item{n}{non-negative integer (vector), for
    \code{factorialZ}.  For \code{chooseZ}, may be a \code{bigz} big
    integer, also negative.}

\item{k}{non-negative integer vector.}
}
\description{
Used for better integer accuracy over non 
negative integer input than base::choose(n,k)
}
