% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.valid.combination}
\alias{is.valid.combination}
\title{is.valid.combination}
\usage{
is.valid.combination(x, n)
}
\arguments{
\item{x}{A vector of combination members}

\item{n}{an integer, n of the combination}
}
\description{
is.valid.combination
}
\examples{
n<-20
# non integer
try(is.valid.combination(0.5,n))
# values out of range
try(is.valid.combination(c(0,n+1),n))
# duplicate values
try(is.valid.combination(c(1,1),n))
# valid
# no argument
is.valid.combination(vector(),n)
# single argument
is.valid.combination(c(1,n),n)
# no argument
is.valid.combination(matrix(c(1,1,n,n),2),n)
}
