% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{print.combo}
\alias{print.combo}
\title{print.combo}
\usage{
\method{print}{combo}(x, ...)
}
\arguments{
\item{x}{a combo object}

\item{...}{ignored}
}
\description{
print.combo
}
\examples{
n<-20
k<-10
combo<-createCombo(ceiling(runif(100)*choose(n,k)),n,k)
print(combo)
}

