% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revCombnG.R
\name{revCombnG}
\alias{revCombnG}
\title{Reverse Combination Generator}
\usage{
revCombnG(x, n)
}
\arguments{
\item{x}{integer, a single combination as a 'vector' or several (with the same k) as rows in a 'matrix'}

\item{n}{integer, length(1), <= .Machine$integer.max, size of the set from which combination is generated}
}
\value{
a 'vector' of the requested indices.
}
\description{
Returns the reference integer vector of a combination matrix\cr
k is identified from length(x) or length(x[1,]) as appropriate.
}
\examples{
n<-15
l<-revCombnG(rbind(
c(2,7),
c(3,5),
c(3,9),
c(10,1),
c(6,3),
c(5,2)
),n)
print(l)
k<-2
combnG(l,n,k)
}
\author{
Dale Potter, \email{dale@piratepress.org}
}
