% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{sampleCombo}
\alias{sampleCombo}
\title{sampleCombo}
\usage{
sampleCombo(n, k, sampleSize = 1000)
}
\arguments{
\item{n}{non-negative integer (vector), for
    \code{factorialZ}.  For \code{chooseZ}, may be a \code{bigz} big
    integer, also negative.}

\item{k}{non-negative integer vector.}

\item{sampleSize}{an integer defining the size of sample to draw. Default 1000.}
}
\description{
Given n,k,sampleSize, produce an ultraCombo object of that
n,k loaded with samples defined by sampleSize.
}
