% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{slice.combo}
\alias{slice.combo}
\title{slice.combo}
\usage{
slice.combo(combo, slice)
}
\arguments{
\item{combo}{a combo object}

\item{slice}{index of the desired subset of combo object}
}
\description{
slice.combo
}
\examples{
n<-4
k<-2
combo<-createCombo(seq(choose(n,k)),n,k)
slice.combo(combo,seq(3))
slice.combo(combo,seq(3)*2)
try(slice.combo(combo,0))
try(slice.combo(combo,7))
}

