% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{ultraCombo}
\alias{ultraCombo}
\title{ultraCombo}
\usage{
ultraCombo(i, n, k)
}
\arguments{
\item{i}{A vector of indices}

\item{n}{an integer, n of combination}

\item{k}{an integer, k of combination}
}
\value{
A list with the following elements:\cr
$i The indices of the combination set.\cr
$n The n of the combination set.\cr
$k The k of the combination set.\cr
$Gen A function which generates combinations from the indices
}
\description{
Create a combination object.
}
\examples{
n<-20
k<-10
combo<-ultraCombo(seq(choose(n,k)),n,k)
print(combo)
object.size(combo$Gen(seq(combo$len)))
}
