% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.ultraCombo.R
\name{chunk.combo}
\alias{chunk.combo}
\title{chunk.combo}
\usage{
\method{chunk}{combo}(combo, chunkSize)
}
\arguments{
\item{combo}{an 'ultraCombo' to chunk}

\item{chunkSize}{maximum number of indices per chunk.}
}
\description{
chunk.combo
}
