% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{createCombo}
\alias{createCombo}
\title{createCombo}
\usage{
createCombo(i, n, k)
}
\arguments{
\item{i}{A vector of indices}

\item{n}{an integer, n of combination}

\item{k}{an integer, k of combination}
}
\value{
A list with the following elements:\cr
	$i The indices of the combination set.\cr
	$n The n of the combination set.\cr
	$k The k of the combination set.\cr
	$Gen A function which generates combinations from the indices
}
\description{
Create a combination object.
}
\examples{
n1<-10
k1<-5
n2<-choose(n1,k1)
k2<-5
i<-ceiling(runif(1)*choose(n2,k2))
combo<-createCombo(combnGen::combnG(i,n2,k2),n1,k1)
combo
combo$Gen(seq(combo$len))
}

